chore(test): reorganize test mocks
Move to dedicated folder under test.
